requirements.txt
appdirs==1.4.3
argh==0.26.2
blinker==1.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Resize==2.0.3
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
humanize==0.5.1
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysql==0.0.1
MySQL-python==1.2.5
mysqlclient==1.3.4
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.1
pyparsing==2.2.0
semantic-version==2.6.0
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.2
WTForms==2.1
wtforms-recaptcha==0.3.2